pyproject.toml
[tool.poetry]
name = "bot-shop-v1"
version = "0.1.0"
description = ""
authors = ["papayni <mark.uzun7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tortoise-orm = "^0.20.0"
aerich = "^0.7.2"
pydantic-settings = "^2.1.0"
aiohttp = "^3.9.1"
coloredlogs = "^15.0.1"
aiogram-dialog = "^2.1.0"
uvicorn = "^0.24.0.post1"
uvloop = "^0.19.0"


[tool.pyright]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
